refactored from yaml classmethod into user object
